requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
attrs==25.3.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
diff-match-patch==20241021
distro==1.9.0
Django==5.2.7
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.2
django-guardian==3.2.0
django-import-export==4.3.12
django-modeltranslation==0.19.17
django-silk==5.4.3
django-simple-history==3.10.1
django-timezone-field==7.1
django-unfold==0.69.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
drf-yasg==1.21.11
environs==14.3.0
gprof2dot==2025.4.14
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
lxml==6.0.2
marshmallow==4.0.1
openai==2.1.0
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tablib==3.9.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
gunicorn==20.1.0
django-celery-results==2.6.0